requirements.txt
audoai-common==0.3.1
audoai-noise-removal==1.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ffmpeg-python==0.2.0
filelock==3.18.0
fsspec==2025.5.1
future==1.0.0
hf-xet==1.1.5
huggingface-hub==0.14.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
py-real-esrgan==2.0.0
pycparser==2.22
pyheif==0.8.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
shiboken6==6.9.1
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
urllib3==2.5.0
websocket-client==1.8.0
